changed package name to net.mf added tags added some advanced data passing for scenarios
