--- src/test/java/net/hpe/MyStepdefs.java
@@ -1,27 +1,0 @@
-package net.hpe;
-
-import cucumber.api.PendingException;
-import cucumber.api.java.en.Given;
-import cucumber.api.java.en.Then;
-import cucumber.api.java.en.When;
-
-public class MyStepdefs {
-    @Given("^I have A$")
-    public void iHaveA() throws Throwable {
-        // Write code here that turns the phrase above into concrete actions
-        //throw new PendingException();
-    }
-
-    @When("^I add B$")
-    public void iAddB() throws Throwable {
-        // Write code here that turns the phrase above into concrete actions
-        //throw new PendingException();
-    }
-
-    @Then("^I have A and B$")
-    public void iHaveAAndB() throws Throwable {
-        // Write code here that turns the phrase above into concrete actions
-        //throw new PendingException();
-    }
-
-}
